add claims

+++ services/data/users/src/users.rs
@@ -273,7 +273,7 @@ impl Users {
         &self,
         user_id: Uuid,
         tenant_id: Uuid
-    ) -> Result<Vec<(i8, String)>, String> {
+    ) -> Result<Vec<(i64, String)>, String> {
         info!("Users::get_user_permissions()");
 
         let result_stmt = self.client.prepare_cached(
@@ -292,7 +292,7 @@ impl Users {
                     Ok(rows) => {
                         let mut v = Vec::new();
                         for r in rows {
-                            let permission_id: i8 = r.get("permission_id");
+                            let permission_id: i64 = r.get("permission_id");
                             let permission_name: String = r.get("permission_name");
 
                             v.push((permission_id, permission_name));
